Use symbolic values for segment/object first/none

+++ common/main/laser.h
@@ -164,7 +164,7 @@ int laser_are_related(int o1, int o2);
 extern int do_laser_firing_player(void);
 extern void do_missile_firing(int drop_bomb);
 extern void net_missile_firing(int player, int weapon, int flags);
-extern int Network_laser_track;
+extern short Network_laser_track;
 
 int Laser_create_new(vms_vector * direction, vms_vector * position, int segnum, int parent, enum weapon_type_t type, int make_sound);
